intel_idle: S0ix are not always the real C states of CPU
BZ: 142967

On some platforms S0ix are not the real C states enumerated
by CPU and hence this check is not required anymore. Fix it.

Also, since platforms like BYT are no longer intel-mid, but
follows the same strategy, let's fix this platform as well.
One can debate SOC_POWER can be enabled for BYT, but that
config has a tight relation with PMU which BYT is no more
interested in it. So, rather enabling ATOM_SOC_POWER for
BYT, let's use the runtime checks which will help enabling
the single binary support for all these patforms in future.

Change-Id: I5084122195d87844bf96ebec1dca0457ee2741d7
Signed-off-by: Srinidhi Kasagar <srinidhi.kasagar@intel.com>
Signed-off-by: Vishwesh M Rudramuni <vishwesh.m.rudramuni@intel.com>
Reviewed-on: http://android.intel.com:8080/137988
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Tested-by: Roux, Nicolas <nicolas.roux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>