[WORKAROUND]Audio: PCM: Add codec_active check in close_delayed_work
BZ: 143872

This is the workaround for the issue seen in saltbay where when music is
played and then when voip call is made and disconnected, media resumes
but then after few seconds the audio is not heard, that is becasue of
the pop_wait update that happens based on the rtd->pop_wait variable which
replaces the codec_dai->pop_wait used in 3.4.
Codec_active field is also checked along with the pop_wait to handle scenarios
where different cpu_dai_link can map  to same codec_dai_link.
This should be fixed with DPCM when this patch can be reverted.

Change-Id: If6902477ddb210be5818461215778bb4e4976d3a
Signed-off-by: Mythri P K <mythri.p.k@intel.com>
Reviewed-on: http://android.intel.com:8080/138562
Reviewed-by: Kp, Jeeja <jeeja.kp@intel.com>
Reviewed-by: B, Jayachandran <jayachandran.b@intel.com>
Reviewed-by: R, Dharageswari <dharageswari.r@intel.com>
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>