audio: add mixers to the model
BZ: 148845

SET_SWM command is sent to set the mixers in the FW.

Each mixer control input is mapped to one bit of the 32-bit mixer register.
Setting this mixer control via userspace causes the corresponding bit to be set
in the register. When DAPM is triggered, the SET_SWM command is generated by
looking at all the bits set for each mixer and finding their corresponding
PATH_ID from the lookup table.

Change-Id: Ic2dad27c2c49ffd3dc01b01bc810075212a5626b
Signed-off-by: Omair Mohammed Abdullah <omair.m.abdullah@linux.intel.com>