[BACKPORT]x86: Remove WARN_ON_ONCE(in_nmi) from vmalloc_fault
BZ: 143851

Since the NMI iretq nesting has been fixed, there's no reason that
an NMI handler can not take a page fault for vmalloc'd code. No
locks are taken in that code path, and the software now handles nested
NMIs when the fault-enables NMIs on iretq.

Not only that, if the vmalloc_fault() WARN_ON_ONCE() is hit, and
that warn on triggers a vmalloc fault for some reason, then we can go
into an infinite loop(the WARN_ON_ONCE() does the WARN() before updating
the variable to make it happen "once")

Reported-by: "Liu, Chuansheng" <chuansheng.liu@intel.com>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>

This patch get from the community:
Link: https://lkml.org/lkml/2013/10/15/894

Change-Id: Ifa493df1abdc7b05309a8534515284ef6687bb5c
Signed-off-by: lizhuangzhi <zhuangzhi.li@intel.com>
Reviewed-on: http://android.intel.com:8080/138776
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: Liu, Chuansheng <chuansheng.liu@intel.com>
Tested-by: Ng, Cheon-woei <cheon-woei.ng@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>