Audio: wm8958: Set MICBIAS high during jack insert/remove for micd_en_delay time
BZ: 124596

Currently, micd_en_delay is set to 5sec. For this long the MICBIAS will be
held high during jack insertion/removal as this would prevent the breaking
sound heard during the partial insertion of jack.

The breaking sound was heard due to the MICBIAS pulsing.

The constraint currently is 5sec delay for partial insertion due avoid
hearing breaking sound, if the user takes longer to insert the breaking
sound would still be heard until full insertion of jack.

Change-Id: I51afdc82181cd57deb2b4d02c361cc84f7288505
Signed-off-by: Samreen Nilofer <samreen.nilofer@intel.com>
Signed-off-by: Paul Gomme <Paul.Gomme@Wolfsonmicro.com>
Reviewed-on: http://android.intel.com:8080/139495
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Reviewed-by: Shah, Hardik T <hardik.t.shah@intel.com>
Reviewed-by: Diwakar, Praveen <praveen.diwakar@intel.com>
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Reviewed-by: Abdullah, Omair M <omair.m.abdullah@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Reviewed-by: B, Jayachandran <jayachandran.b@intel.com>
Tested-by: Koyili, SrikanthX <srikanthx.koyili@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>