drm/i915: Added a new render specific hw(VLV) workaround 'WaSendDummy3dPrimitveAfterSetContext'
BZ: 154720

This workaround is needed for the HW context feature
It is used after adding the mi_set_context command in
ring buffer for Hw context switch.
As per the Bspec "The software must send a pipe_control
with a CS stall and a post sync operation and then a
dummy DRAW after every MI_SET_CONTEXT and after any
PIPELINE_SELECT that is enabling 3D mode.
A dummy draw is a 3DPRIMITIVE command with Indirect Parameter
Enable set to 0, UAV Coherency required set to 0,
Predicate Enable set to 0,End Offset Enable set to 0, and
Vertex Count Per Instance set to 0, All other parameters are don't care"

Change-Id: Ida785457d2108a31853ba1de9b8da94acebcef73
Signed-off-by: Akash Goel <akash.goel@intel.com>