EM/THERMAL: Change the order iio map list is built
BZ: 146822

[Description of issue]
The temperature values for different zones reported
from the thermal class were wrong/invalid

[Implemented Solution]
In the iio driver the adc channel map list was getting
built in the reverse order. This was causing the adc values
returned to thermal driver to be in improper order. This
order of building the list is fixed now.

Category: feature differentiation
Domain: EM.ThermalManagement-commmon
Origin: internal
Upstream-Candidate: yes

Change-Id: Idd1ba4734193bfa9e0b4c5397e737ee6872374e1
Signed-off-by: Ajay <ajay.thomas.david.rajamanickam@intel.com>
Reviewed-on: http://android.intel.com:8080/140231
Reviewed-by: Pawnikar, Sumeet R <sumeet.r.pawnikar@intel.com>
Reviewed-by: B M, Shravan K <shravan.k.b.m@intel.com>
Reviewed-by: Tc, Jenny <jenny.tc@intel.com>
Tested-by: Koosalapur, Siddalingesh S <siddalingesh.s.koosalapur@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>
Reviewed-by: Dower, Robert <robert.dower@intel.com>
Tested-by: Dower, Robert <robert.dower@intel.com>