[PORT FROM R42B-STABLE] drm/i915: fix MIPI blankout when playing extended video
BZ: 139642

When playing extended video with MIPI and HDMI, When the mouse is moved,
video player control should appear on the MIPI display. But, MIPI display
is blanking out. The issue is that bandgap reset was not done in dpms on
sequence. Bandgap reset was done only in full mode set sequence. Now
moving the bandgap sequence code to dpms on/off code flow path too.

Change-Id: Ib208ea87d74ed0f51836aba7040eeea8e0f3e330
Signed-off-by: Yogesh Mohan Marimuthu <yogesh.mohan.marimuthu@intel.com>
Reviewed-on: http://android.intel.com:8080/133102
Reviewed-by: Kumar, Shobhit <shobhit.kumar@intel.com>
Reviewed-by: Yang, Fei <fei.yang@intel.com>
Tested-by: Sun, Hang L <hang.l.sun@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>