audio: add CODEC-CODEC link for modelling Modem
BZ: 148845

This is a link from the CODEC BE to the MODEM BE. The codec hw_params() are
called whenever DAPM detects a valid route between the CODEC-MODEM without the
need for any stream to be active.

A valid route is formed between the two by routing the SBA mixers appropriately.
This link is not visible to userspace even though it is present in the dai_link.

This link is also triggered when there is a route between the CODEC and any DAPM
INPUT/OUTPUT. So it will be triggered in the Aware/VAD case as well.

Change-Id: I306c2865e8b1ee80e50db1518bce9c6e52ac07e0
Signed-off-by: Omair Mohammed Abdullah <omair.m.abdullah@linux.intel.com>