regmap: fix: read the mask registers again to see if the interrupts are masked
BZ: 141086

FIFO errors were seen in Merrifield even though we had the this FIFO IRQ masked.

Issue:
regmap_irq_thread() would trigger the FIFO irq whenever it reads
the status register when it receives other IRQs from the codec.

Rootcause:
This was due the masked register in the regmap not having the correct value.
So, whenever the status register of the codec was read if the fifo err status
bit is also set with the other status bits the regmap_irq_thread() would trigger
fifo isr along with other isrs.

The mask bit in the codec hw register was updated correctly but
the mask val in the regmap was incorrect.

This patch aims at fixing the SW logic issue, not a HW issue.

Change-Id: I5747630eb128d37b6f1f988f267c6ffc1bd95230
Signed-off-by: Samreen Nilofer <samreen.nilofer@intel.com>