CONFIG: Enable NETFILTER_XT_MATCH_QUOTA2_LOG
BZ: 146075

For Android, this option is needed for data usage warning to work.
Also some platforms are missing MULTIPLE_TABLES, this patch enables
them. Running "make menuconfig" introduces some other changes that
are not the intention of this patch, but already enabled by other
patches which didn't update defconfig accordingly.

Change-Id: Ia272884289a34f7bc03b270da7217d4ce653e084
Signed-off-by: Fei Yang <fei.yang@intel.com>