[PORT FROM K3.4] legacysuspend: native_cpu_die sequence enhancement in cpu hotplug
BZ: 100101

In order to end a logical unplug of a cpu, a percpu state CPU_DEAD is set
by the targeted cpu's idle task. In a racing case (~every time), this
value is being set after being checked, which was causing a 100ms retry delay.

To avoid this racing scenario, we replace the looping/retry check by a
completion structure. This makes the end of the unplug sequence more reliable
and safely applicable to all platforms.

In a legacy Suspend-To-RAM (S3), sequence is then shorten by ~300ms on CLV
(3 nonboot cpus) which leads to savings in all use cases where S3 is involved
(~0.3mW in Standby-Flight Mode).

Orig-Change-Id: I8ebb58e1482cc72d6f90ef98cc200179398fca08
Change-Id: Idc62cd52e7e593a5985e3baa15f56b13aeccc9af
Signed-off-by: Anthony Toubeau <anthony.toubeau@intel.com>
Signed-off-by: Fernand Cuesta <fernand.cuesta@intel.com>