drm/i915: Fix for debugfs/sysfs use of snprintf
BZ: 149012

Modified the following construct:

	len = snprintf(buf, sizeof(buf), ...);

	if (len > sizeof(buf))
		len = sizeof(buf);

snprintf returns the number of characters (excluding the NULL)
that *would have been* written if the buffer was large enough,
so it can exceed sizeof(buf). The test should really be:

	if (len >= sizeof(buf))
		len = sizeof(buf) -1;

However, there is a function available that already includes
this test "scnprintf()" so use that instead.

Also modified two instances of sprintf to use scnprintf in i915_debugfs.c.

Change-Id: I8c9a90a2724351f0b7bab011e67d1751dd5d7027
Signed-off-by: ian-lister <ian.lister@intel.com>