[PORT FROM K3.4] hsic: Add 25ms wake lock in Aux 2 ISR to prevent hsic to go in low power before hsic resume
BZ: 124539

We detected Aux2 request was not always able to wake up the HSIC link.
Aux 2 is high. ISR is executed. link resume is done but just after link suspend is done.
It is needed to prevent that, for instance by taking a wakelock during the 20ms of the bus resume.

Change-Id: Iad0f3703e92a13be0a00a2d20e235d2289166799
Signed-off-by: feng wang <feng.a.wang@intel.com>