[PORT FROM K3.4] usb/keyboard: remove leds set/unset for usb keyboard in driver/kernel
BZ: 103530

In ICS, framework will handle the leds for usb keyboard and handle
related function. In keyboard driver, leds are also set/unset but
the implementation conflicts with that of the framework.
The implementation of framework is same as google/samsung ICS phone.
So remove leds set/unset functon in keyboard driver.
Please note that the implementation of android is different from
other linux base platform like ubuntu.

Change-Id: Iec9aa964a144236b34be62ff8012828779c8b7ee
Signed-off-by: feng wang <feng.a.wang@intel.com>