[PORT FROM R42B-STABLE] Disabled primary plane power optimization
BZ: 146360

This bug is due to one patch (109183) missing during kernel 3.10 upgrade.

Disabled the primary plane power optimzation when the sprite resolution >= the
primary plane resolution, the primary plane is disabled. we use sprite for any
layers other then video with alpha blending enabled, so diabling the primary
with the alpha blending enabled causes lot of flickering issue and the sprite
transparent region is blanked.

Change-Id: I17ce0988af7fca197aff0ad5e203772e17c77e25
Signed-off-by: Lei Zhang <lei.zhang@intel.com>