intel_mid_dma: Reduce the wait time for suspending dma channel
BZ: 149684

Currently the wait time is 10ms. It is a little much for some devices
like BT, as the user space BT stack may think there is some data link
error when no data packet coming for a short while.

And based on lots of tests the dma channel can either be suspended right
away in less than 1ms or can't be suspended at all even after 100ms. So
simply cut it to 4ms. One other change is to set the each delay to 100
us instead of 1ms to improve the granularity.

Also when the dma channel can't be suspended(rarely happen), it's better
to notice user about it.

Change-Id: I0f46cd610db87b885659d8e1e009f5f628d71b91
Signed-off-by: Feng Tang <feng.tang@intel.com>