configs: Merged 64-bit MRFL defconfig from multiple source: baseline
BZ: 147826

This is a first sweep of Merrfield 64bits kernel
configuration to have a stable adb shell console.

In comparison to the 32bits version, it should be noted that
 - atomisp2 was disabled
 - sep54 was disabled

Once the port to a 64bits kernel will be more advanced, the
kernel configuration should be amended.

Create a baseline defconfig for MRFL 64-bit from multiple
source. This baseline boots successfully to shell.

Change-Id: I83ba4c7bf1472e4fa55bbbd2d9b97d7f12325d30
Signed-off-by: Raj Mojumder <raj.mojumder@intel.com>
Signed-off-by: Patrick.Noziska <patrick.noziska@intel.com>
Signed-off-by: Robert Jarzmik <robert.jarzmik@intel.com>
Signed-off-by: Guilhem IMBERTON <guilhem.imberton@intel.com>