EM/charging_framework: Validate the cur_state setting from user space
BZ: 140965

[Issue Description]
cur_state setting in the cooling device user interface was not
checking the validity of the values provided from user space.
As a result, invalid value setting does not throw an error back
to the user.

[Implemented solution]
The value received from user space is validated against the num of
states allowed and if it is beyond the limit, error is thrown back
to the user.

Category: Device Enablement
Domain: EM.Battery-Charging Framework

Origin: internal
Upstream candidate: yes

Change-Id: I0d3443b09c169a449409445a703d4f766a480b0b
Signed-off-by: yegnesh <yegnesh.s.iyer@intel.com>