[PORT FROM K3.4] Trigger reselection for wkup caused by other CPU
BZ: 119142

On CLV based devices due to independent C6, when all CPUs are in
deep idle state (C2 and above) an interrupt will wakeup all CPUs
even if it was adressed to only one. In this case, for the victim CPUs
do the menu selection again in order to choose the right c-state
based on the remaining idle time. However in the reselection process
(menu_select) consider a perfect prediction when updating the buckets,
as the CPU prediction was not wrong it was wokenup by another one.

With this patch a power gain of ~3% is seen on video record UC as
well as browsing use case

Change-Id: Ia0181d92b7305adf759139907b48573aa8616902
Orig-Change-Id: If214e2a9f019b020d2ba93724e3c4e79fe20369d
Signed-off-by: Yuyang Du <yuyang.du@intel.com>
Signed-off-by: nchalhou <nicole.chalhoub@intel.com>
Signed-off-by: Anthony Toubeau <anthony.toubeau@intel.com>
Signed-off-by: Illyas Mansoor <illyas.mansoor@intel.com>