drm/i915: Modifying RC6 Promotion timer for Media workloads on VLV
BZ: 155751

In normal cases, RC6 promotion timer is 1700us. This will result in more
time spent in C1 state. For more residency in C6 in case of media workloads,
this is changed to 250us. Not doing this for 3D workloads as too many C6-C0
transition delays can result in performance impact. Tracking the media
workloads based on command submission to MFX engine

Change-Id: I82dcd90689b1bba5a37c92a94f7be38bea80d2a1
Signed-off-by: Satyanantha, Rama Gopal M <rama.gopal.m.satyanantha@intel.com>