kernel config: cleanup - run scripts/updatedefconfigs.sh to set baseline state
BZ: 141353

This change is the result of running scripts/updatedefconfigs.sh to
establish a baseline of default kernel configs for subsequent commits.
It has no other changes.

This allows the deltas in dependent changes to result from only their
real changes and not from pseudo changes arising from updates to the
kernel config from parents of this commit.  This would not be an issue
if scripts/updatedefconfigs.sh was uniformly executed as a part of
the process anytime that the kernel default config files are modified.

Change-Id: Ide6bb0cd5b84420418f9811ca7588636f1924b23
Signed-off-by: Dale Stimson <dale.b.stimson@intel.com>