[Telephony/MCD]: Add Cherryview CPU and PMIC_CHT
BZ: 150172

In order to get data form ACPI table for Cherrytrail
platform, we need to add the CHERRYVIEW CPU and manage
the PMIC for cherrytrail. The modifications are applicable
for ACPI scope only.
Because the telephony configuration name is added in the
ACPI table, it can be directly stored in the platform
data. This change must be backward compatible with the
oldest table where the configuration is not available in
this table.

Change-Id: I34aeebffca60f576f317439922de92cf23a9b000
Signed-off-by: Gwenn Bourree <gwenn.bourree@intel.com>