audio: add probe support to dpcm
BZ: 144797

Adds probe point support. There are 8 simultaneous probe streams possible, with
each of their connection points defined using ALSA controls. The probe DAI is
opened with the correct substream (for which the ALSA control has been
programmed).

A _PROBE command is sent for that probe pipe after the ALLOC has been sent.
The probe pipe_id is filled from the stream map into the _PROBE command.

Only gain endpoints are supported currently for all pipelines. A stream should
be running in that pipeline to probe any point in the pipeline.

Change-Id: I9c50c4171996eb90a76d9be55db29e61f673ee0d
Signed-off-by: Omair Mohammed Abdullah <omair.m.abdullah@linux.intel.com>