[PORT FROM R42B] drm/i915: invalidate tlb for hung ring in reset path
BZ:152028

Summary:
During MTBF test UIWDT issue is seen. This is rootcaused to HW Status
page not getting updated with the latest seqno, causing ring reset.
Adding TLB invalidate in ring reset path as well to make sure we are
not again reading a stale address from the TLB after this reset.

Change-Id: Ibb0f9650e7219968da2d0c26e61b2a74045779e3
Signed-off-by: Naresh Kumar Kachhi <naresh.kumar.kachhi@intel.com>