Audio: Update locking mechanism in sst irq thread
BZ: 146566

As threaded IRQs can run on both the CPUs (CPU0 and CPU1),
there may be possibility that CPUs are acting on same msg at
a time.
spinlock is taken for "list_for_each_entry_safe" & "list_del"
function. This ensures that current msg node is deleted from
the list, so that if interrupt comes in other CPU, it will get
new msg from the list.

Change-Id: I6763e9d1894ce2b2fc62cbdeddd6a6660d1fd543
Signed-off-by: Praveen Diwakar <praveen.diwakar@intel.com>