EM/charging_framework: Charge current should not be throttled when charger is not connected
BZ: 146844

[Issue Description]
When charger is not connected and the thermal zone moves to warning state,
charge current was throttled.

[Implemented solution]
When charger is not connected, charge current should not be throttled.
Hence, if charger is not connected, return without throttling.

Category: Device Enablement
Domain: EM.Battery-Charging Framework
Origin: internal
Upstream candidate: yes

Change-Id: Ib0d5ded9482571d159986cdc8ec588fb64fbac77
Signed-off-by: rjayapax <ranjinix.jayapal@intel.com>