audio: add aware support to dpcm
BZ: 144797

The aware module is triggered by configuring the input and then closing the
aware switch. This causes the Aware algorithm to run with the preconfigured
aware classifier parameters. Opening the switch causes the link to break and the
algorithm stops.

This is modeled in DPCM using a CODEC<->CODEC link, with Aware being and
internal sink of the FW. There is no data transferred to/from IA. The codec is
triggered on by configuring the codec and platform paths and then closing the
aware switch.

Change-Id: I555ae2b90e8598e3cb2e89bf938a5fdc58995f74
Signed-off-by: Omair Mohammed Abdullah <omair.m.abdullah@linux.intel.com>