audio: add aware support to dpcm
BZ: 167991

The aware module is triggered by configuring the input and then closing the
aware switch. This causes the Aware algorithm to run with the preconfigured
aware classifier parameters. Opening the switch causes the link to break and the
algorithm stops.

This is modeled in DPCM using a CODEC<->CODEC link, with Aware being and
internal sink of the FW. There is no data transferred to/from IA. The codec is
triggered on by configuring the codec and platform paths and then closing the
aware switch.

Change-Id: Ibe92f69594ce3bb0bf5fd4fb002481083db7a5a7
Signed-off-by: Omair Mohammed Abdullah <omair.m.abdullah@intel.com>