gfx/rgx: Add kernel config CONFIG_GFX_RGX_BVNC for IMG DDK 1.3
BZ: 158390

Add kernel config CONFIG_GFX_RGX_BVNC and normalize kernel config
files.  The BVNC value must be the same for IMG kernel mode and user
mode drivers.

Current values (as of this time) are:
- Merrifield - 1.14.4.4 - DDK1.2
- Merrifield - 1.76.4.6 - DDK1.3
- Moorefield - 1.72.4.12 - either DDK 1.2 or 1.3

This is being done as part of new support for Moorefield, as the
Merrifield and Moorefield BVNC values are distinct.

Change-Id: Ib273cd4b38e53ad130c64c66ca8c9a6c023119a9
Signed-off-by: Dale Stimson <dale.b.stimson@intel.com>