[BYT/DMA] do not wait 100usecs for FIFO to be empty
BZ: 155406

In 'dma_wait_for_suspend' do not wait for a minimum
of 100usecs for the FIFO to be empty. Most of the time,
the FIFO is emptied after a 'cpu_relax'.
If the FIFO is not emptied after a 'cpu_relax', retry
with an incremental backoff mechanism and give up
if the FIFO is not emptied after 100 loops which
is about 5ms.
This will hopefully increase the speed of Bluetooth OPP
transfers, which are lower on BYT compared to other
platforms.

Change-Id: I17246ccfe71fe83a4a547e5a1a5b28029d6d7e64
Signed-off-by: Frode Isaksen <frodex.isaksen@intel.com>