Audio: Set Media_loop format based on board config
BZ: 144797

Media_loop needs to be configured based on the Speaker/DMIC
configuration. on MRFLD, the Speaker is Mono and DMIC is
Stereo. based on the format, the media loop EQ params
needs to set.

added format to the widget private data for media_loop.
for MRFLD, media_loop will be configures as MONO and
media_loop2 will be configured as Stereo. This will be filled
by firmware config.

Change-Id: I6d5a99da69bfcc1088c0b56b7ce806cfe5a03cf1
Signed-off-by: Jeeja KP <jeeja.kp@intel.com>