SSP SPI frequency isn't equal expect value on MRFL
BZ:149006
When SSP work in SPI mode on MRFL, the output frequency
isn't equal the expect value.
It is equal 4 times the expect value.
So, update the reference clock when calculate.

And due to orignal reset SSP code is write fix number into
frequency register, so update it to write the right data
after calculate frequency.

Change-Id: I0d8b02095bb3aa13cfe0fddb6bc01b8f097a6d34
Signed-off-by: Eric Xiong <eric.xiong@intel.com>