drm/i915: Resolved the ColorConfig Gamma Stress issue
BZ: 155211

Resolve ColorConfig -gamma stress issue. It's a corner case for
color manager implementation while applying gamma on planes. Due
to a typo error, when gamma correction was enabled on PIPE A,
sprite B was not getting updated, effects were seen only on sprite A.
Changed the MACRO to reflect correct Sprite plane.

Change-Id: I6be0fdf69cb008d22811df840a7dcc7c32428396
Signed-off-by: Sardar Shamsher Singh <sardarx.s.singh@intel.com>