ASoC: core: call mute/unmute for compress stream CPU Dai
BZ: 144797

For Compress stream, digital mute can be applied
before start and stop of stream to minimise
any system noise.

This patch adds soc_digital_dai_mute() is called by the core for
each cpu DAI when the mute is applied or freed. Stream will
be umuted before start stream and muted when stream is closed.

Change-Id: Ifdb28dc8a6d98143c86d8ff408268e5b85d8d64d
Signed-off-by: Jeeja KP <jeeja.kp@intel.com>