[REVERT ME] xhci: Add 1ms delay before stop endpoint cmd
BZ: 142129

In some situation was observed that after the stopping
of an endpoint the event: stop - length invalid is
received from the xHCI controller. Following this event
the endpoint hangs in this state resulting in an
abnormal data transfer.
During the test was observed that adding 1ms between
the stop endpoint commands is fixing the issue.
This is a temporary solution until the root cause of
the issue is identified.

Change-Id: I10358ba5784ea90c7ed5a1d0f94cdb9f8139c99e
Signed-off-by: George Sandu <georgex.sandu@intel.com>