drm/i915: Drop forcewake w/a for missed interrupts/seqno on Sandybridge
BZ: 160545

[Port from 3.13.0]
commit 67e5871be82fec1451801d448b51d9a403d1ffac
Author: Chris Wilson <chris@chris-wilson.co.uk>
Date:   Fri Nov 22 20:35:24 2013 +0000

I believe, and an evening of i-g-t, that our original workaround for the
missed interrupts on Sandybridge, that of holding forcewake whilst we
wait for an interrupts, is no longer required. This leaves us dependent
on the second workaround of forcing an UC read of the ACTHD before
reading back the seqno from the snooped HWS. Dropping the forcewake
should allow us to conserve a little power, not much as the GPU is meant
to be busy whilst we wait for it!

Change-Id: I938835b9df8b5453ff3a9c5debbec654c5f9076f
Signed-off-by: Deepak S <deepak.s@intel.com>
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>