[Telephony/MDM_CTRL]: update platform_data to manage configuration name from SFI table
BZ: 157236

Because the IFWI will directly provide the telephony
configuration name to platform data, but the oldest IFWI for
the other platforms still provide modem names, the modem control
platform data SFI table handle must be updated. Then when the
configuration name is provided, the modem name is extracted from
this name and the configuration is given as it is. In case the
modem name is extracted from the SFI table, the configuration
is generated from a lookup table.
For HSI platform data, the modem association table is modified
to support both ways of coding the modem name.

Change-Id: Ic2fd83e627819c19aa3fe8526f96f89b6f5620b1
Signed-off-by: Gwenn Bourree <gwenn.bourree@intel.com>