Audio: Fix to put LPE in D0i3 after boot for the ACPI devices
BZ: 160260

LPE device remains on D0 after boot. The initial runtime PM status of
LPE devices is 'suspended', but it need not reflect the actual physical
state of the device. Thus, if the device is initially active, its runtime
PM status must be changed to 'active', with the help of
pm_runtime_set_active(), before pm_runtime_enable() is called for the device.

Change-Id: I1fa162e588da7dadeb13e553a4f5951099029b82
Signed-off-by: Dharageswari.R <dharageswari.r@intel.com>