[Revert Me] SPI: add SSP ACPI platform driver probe to unblock PM
BZ: 148847

Provide SSP ACPI platform driver probe to unblock PM/RTPM validation on CHT.

Note, we don't do any actual initialization in the probe except enabling the
PM of the device. This is a temporary patch needs to be reverted later
when the new SSP driver is merged to mainline.

Change-Id: I5d25e695eded65a7b31e804ba43fbe8b2a6595e5
Signed-off-by: Jason Chen <jason.cj.chen@intel.com>