mei: use driver's power gating state to block writing
BZ: 158216

Since there is not complete sync between runtime pm
and internal power gating states we cannot relay on runtime
pm state to check we can access the host buffer,
but rather we check for the power gating state.

We add hbuf_aquire wrapper which checks all condition
for accessing the hbuf and it also prevent from
other flows to access hbuf by unsetting hbuf_is_ready.

This fixes glitch in txe when we can miss write
because we're depending on aliveness interrupt to trigger
delayed write requests. Aliveness interrupt
kicks in earlier then the pm framework resume handler
completes. In some scenarios pm request can be already in
progress and device might be alive so we cannot
queue the request since we won't have any trigger that
will kick the transfer.

Change-Id: I6e252a7b8ef3c3c10c808f3a3f417625ca650704
Signed-off-by: Tomas Winkler <tomas.winkler@intel.com>
Signed-off-by: Alexander Usyskin <alexander.usyskin@intel.com>
Reviewed-on: https://git-isdc.jer.intel.com/gerrit/1331