intel_soc_pmu: add d0i0 residency stats
BZ: 158555

In current mid_pmu_states dump we do not get d0ix residency and time
on Merrifield.

So propose to add d0i0 time and residency information.

Also this patch removes reg, and mask info which was not very useful
to make sure the data fits in one screen window.

Change-Id: Ifcf48827fc647b85311288f36185f907ee03b4ad
Signed-off-by: Illyas Mansoor <illyas.mansoor@intel.com>