EM/THERMAL: Do not read cached temperature in IRQ handler
BZ: 146185

The update_temp() method in the coretemp driver returns
the previously cached temperature value if it is read again
within a second. This method is used in IRQ handler, where the
driver sends the current temperature as part of the UEvent
to user space. Sometimes, this read (due to very frequent
interrupts) falls within one second and hence returns a
stale value.

This patch avoids this behavior, by directly reading the
temperature from 0x19C, _everytime_ when there is an
interrupt. While we are there:
 * Remove update_temp() as that code is now used
   within show_temp() itself
 * Print threshold0/1 as part of log message
 * Send UEvent only when we go below lower threshold
   or go above upper thresholds (considered as
   valid events)
 * Print log messages only for valid events
 * For invalid events, include the prints as
   pr_debug so that they appear only during debug.

Change-Id: Ib4f32cd8d50ecae9ae18eb41768262b25df770a2
Signed-off-by: Durgadoss R <durgadoss.r@intel.com>