basin_cove_gpadc: Correcting the iio pointer definitions to avoid warnings
BZ: 158068

Issue description: The iio_channels pointers in the adc structure was defined
as pointers to auto variables and it was made to point to const variables.
 This caused compilation warnings.

Solution: Corrected the iio_channels pointer variable type to const to resolve the warnings.

Change-Id: Ie2ffdc46152c534a577fc667e224837eea923d87
Category: device enablement
Domain: BatteryManagement-Common
Origin: internal
Upstream-Candidate: no, not aosp
Signed-off-by: yegnesh <yegnesh.s.iyer@intel.com>