[REVERTME] checkpatch: report COMPLEX MACROS check as warning
BZ: 148845

Report the complex macros check as a warning, so that the patches in this BZ can
be merged.

In this scenario, the check is invalid as complex macros in declarations cannot
be enclosed in do-while or parenthesis.

This patch will be reverted once a scalable way is found to address such
checkpatch exceptions.

Change-Id: I86dccb9b5f16aa3b358d82e0bcc4868ac1d8f8d1
Signed-off-by: Omair Mohammed Abdullah <omair.m.abdullah@linux.intel.com>