atomisp2: Add platform interface for querying camera capabilities
BZ: 158552

This patch adds extended interface for virtual channel support.
get_camera_caps() operation in camera_sensor_platform_data
is used to query the camera capabilities including how many
sensors are connected to this port and how many streams of
each sensor can output.
For legacy sensor drivers in which this operation is not
implemented, the default configuration is one sensor with
one stream output.

Change-Id: Idb0194437a3090b042201bacbd178b9b7d5357c5
Signed-off-by: Tianshu Qiu <tian.shu.qiu@intel.com>