drm/i915: Fix for the kernel panic/crash in the exec-buffer path
BZ: 159029

In the low memory scenario there is a possible case where the
objects which were mapped in GTT at the beginning of the exec-buffer
function could get unmapped, before the batch buffer is dispatched
or they become a part of active list. This is because the shrinker
function could get invoked in the exec buffer path also.
If this happens there could be a kernel crash or the GPU hang.
We see the kernel crash happening due to this issue.

Change-Id: I414510a091e71c7fccc12a0d25b4f756399f88ff
Signed-off-by: Akash Goel <akash.goel@intel.com>