SELINUX: Enable EXT4_FS_SECURITY in x86_64
BZ: 159144

SELINUX requires EXT4_FS_SECURITY to be enabled to support
security labels, otherwise the Kernel won't be able to mount
file systems with SELinux enabled.

Enforcing or permissing mode for SELinux is irrelevant here,
as this will prevent the Kernel to mount the filesystem all-
together indepedently of the SELinux mode.

When this occurs on the /system filesystem, the entire device
will fail to boot properly.

Change-Id: I9e565c25276212361c8ba4b1a8dfe048e25589ec
Signed-off-by: Olivier Fourdan <olivier.fourdan@intel.com>