sdhci: add .setup and cleanup callback for the sdio platform data on MOOR.
BZ: 159258

Wifi needs the GPIO wlan_en to be attached to the vmmc regulator.
This allows the mmc stack to control the power state of the attached
SDIO card. This patch adds the callbacks .setup() and .cleanup()
to the SDIO platform data on Moorefield.

Please note that Moorefield and Merrifield share the same callbacks.
This is not a mistake : the setup procedure is the same.

Change-Id: I2d756a9280c28d9bba9fafef0f958d40d3597fac
Signed-off-by: Aymen Zayet <aymen.zayet@intel.com>