EM/THERMAL: Clean up SoC cooling device in intel_mid_thermal.c
BZ: 158886

The SoC cooling device's cur_state interface does not
return the value that it is set to. It returns different
values depending on the skin thermal state. While this
reflects the functionality, it is very confusing. This
patch modifies the code to return the value that is
echoed, without changing the internal implementation.

The SoC cooling device in TNG/VLV platforms implements
throttling using RAPL whereas the similarly named
cooling device in CLV, implements GFX throttling
using DVFS. So, change the name of this cooling
device from 'SoC' to 'gpu_dvfs' (only for CLV)
to make it close to reality.

Change-Id: Iad18cdfbadeceab6b94925ac98003131415cacc8
Category: feature differentiation
Domain: ThermalManagement-commmon
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Durgadoss R <durgadoss.r@intel.com>