Audio: send media_path and media_loop_map 'OFF' cmd
BZ: 156582

The media loop pipes and media path pipes are switched on
when dapm path is On but the "switch=off" cmd is not
send to tear down the path.
since the tear down of the previous stream is not done, this
will cause the failure in capture stream from next time.

Change-Id: I2b413e6acd2787a8a37810c75697e62bcfdda55e
Signed-off-by: Jeeja KP <jeeja.kp@intel.com>