smb34x-charger: Acquire wakelock for all cable type connect
BZ: 158416

Upon CDP connect the device is going to s3 state but
as the smb347_full_worker is not scheduled
periodically when battery reaches constant voltage
the status from "charging" to "full" is not updated.

This patch acquires wakelock for all type of
cable connect, so that the system will not go to sleep.

Change-Id: I7ded02f25bdd6cc4d074359ea964ad4fc928284b
Category: device enablement
Domain: BatteryManagement-Summit
Origin: internal
Upstream-Candidate: no, needs rework
Signed-off-by: srihars1 <sriharsha.govinda.raj@intel.com>