drm/i915: Change the sprite disable vblank wait to asynchronous wait
BZ: 154194

For static frames rendering it is always better to use a single plane
instead of multiple planes. As part of this switching, we may need to
disable multiple sprites and this incurs a major blocking delay due to
synchrounous vblank waits. Changed the vblank wait and unpin work to be
asynchronous.
e.g., In Android typically to render home screen we use every single
plane possible.. When the image is static, it is better to go for GPU
composition than using sprites to save some power.

Change-Id: I9fe38c8fb2c15563fcb8956392d49259fad9a48b
Signed-off-by: Vijay Purushothaman <vijay.a.purushothaman@intel.com>