Read and store the supp_ia32 version property at boot phase
BZ: 159860

The initial problem was that the ifwi version was not correctly
displayed in POS for SLTB because it used a function specific to
cloverview chip. To unify the way to retrieve versions and avoid
the #ifdef, we now use the system properties.

This patch reads and stores the FTL driver version at boot
phase so it can be read later through
/sys/kernel/fw_update/fw_info/fw_version.
Another patch will read this file at bootime (in .rc files)
and set the matching property.

Change-Id: I640201b40c63dd4e08c4a12d570b72ecbc8b14f2
Signed-off-by: Gaelle Nassiet <gaellex.nassiet@intel.com>