[REVERTME]audio:Do not use VLEX bypass mode for audio
BZ: 153508

During BIAS_OFF we configure the PMIC to use VFLEX
bypass mode. But this is not working in Shadycove PMIC.

This workaround forces to keep the vflex audio rail to
keep on even during audio idle for MOFD platform.

This should be reverted once the issue is root caused.

Change-Id: I01349af5b9c5c279a9c2f54e316f1b7562ca971b
Signed-off-by: Ramesh Babu <Ramesh.Babu@intel.com>