EM/Charger: Disable charger safety timer
BZ: 154634

With load, charging current shall be lower and it will
take more time for the battery to charger to full. If
the safety timer expires, charging will be stopped and
battery shall start discharging. This patch disables the
charger saftery timer, which help to charge continously
as long as the charger is connected.

Change-Id: If69c6e69fba513eea67979bc311ef81bc149ce47
Category: device enablement
Domain: EM.BatteryManagement-Summit
Origin: internal
Upstream-Candidate: no, needs rework
Signed-off-by: R, Kannappan <r.kannappan@intel.com>