pstore/ram: fix bus starvation on ftracing
BZ: 158872

When using pstore as a backend for ftrace in an SMP platform, CPUs are
fighting for making updates to the pstore current index.  On several
CPUs, including latests x86 cores, an atomic cmpxchg on an uncached
memory results in a delay of several us, triggered by the system agent
to prevent a storm of bus locks for direct RAM access.

Therefore, a starvation occurs, where one CPU is always successful in
the atomic_cmpxchg, and the other CPUs are always waiting.

Fix this by moving the hot path (start, size) into a normally cached
memory where no starvation will occur, and where atomic_cmpxchg() will
have a fair distribution. The (start, size), once changed in the cached
area, are copied to the uncached area to still have the correct counters
in the persistent memory.

Moreover, optimize the buffer_start_add() function to use the result of
cmpxchg() and spare one atomic_read() in each loop.

Change-Id: I501c4f683d6976ed1ca58e11f71d112816d9589c
Signed-off-by: Emmanuel Berthier <emmanuel.berthier@intel.com>
Signed-off-by: Robert Jarzmik <robert.jarzmik@intel.com>
Signed-off-by: Emmanuel Berthier <emmanuel.berthier@intel.com>