Kernel: Adding missing Configuration for Advance Netfilter Options
BZ: 161757

The Advance netfilter options are not enabled in 64 bit configuartion.
Hence it prevents to enable the ipt_quota2 module which is required
to have for iptables during the USB Tethering operation enabled.

By enabling this options in the Kernel Config file, the USB tethering
works on X64_BYT_FFRD8.

Change-Id: I379bf510377d660a148d203373d186719332ad25
Signed-off-by: Madhava Rao <sethu.madhavax.rao@intel.com>